Revert "Merge pull request #151 from Rabithua/feat/seo-meta-optimization"
This reverts commit 01ee7d66056eb124f337829313f178c23a72c160, reversing
changes made to b0048c1aa2429e56b99b459c1ce151c0dfb1407b.
