+++ web/src/locales/zh.json
@@ -267,8 +267,7 @@
       "sideBarTitle": "除了这些...还有",
       "helmet": {
         "loading": "加载中...",
-        "defaultDesc": "rote.ink",
-        "profileDesc": "{{name}} 的个人主页"
+        "defaultDesc": "rote.ink"
       }
     },
     "filter": {
