+++ web/src/pages/app/terms/index.tsx
@@ -1,11 +1,8 @@
 import MarkdownDocPage from '@/components/others/MarkdownDocPage';
-import { PageMeta } from '@/components/seo/PageMeta';
 import { useTranslation } from 'react-i18next';
 
 function TermsOfServicePage() {
   const { t } = useTranslation('translation');
-  const pageTitle = t('pages.terms.title', { defaultValue: '服务条款' });
-  const pageDescription = t('pages.terms.loading', { defaultValue: '服务条款页面' });
 
   const buildPath = (lang: string) => {
     const short = lang.slice(0, 2);
@@ -14,14 +11,10 @@ function TermsOfServicePage() {
   };
 
   return (
-    <>
-      <PageMeta title={pageTitle} description={pageDescription} />
-
-      <MarkdownDocPage
-        buildPath={buildPath}
-        loadingText={t('pages.terms.loading', 'Loading terms of service...')}
-      />
-    </>
+    <MarkdownDocPage
+      buildPath={buildPath}
+      loadingText={t('pages.terms.loading', 'Loading terms of service...')}
+    />
   );
 }
