+++ web/src/pages/archived/index.tsx
@@ -3,7 +3,6 @@ import NavBar from '@/components/layout/navBar';
 import TagMap from '@/components/others/tagMap';
 import RandomRote from '@/components/rote/randomRote';
 import RoteList from '@/components/rote/roteList';
-import { PageMeta } from '@/components/seo/PageMeta';
 import ContainerWithSideBar from '@/layout/ContainerWithSideBar';
 import type { ApiGetRotesParams, Rotes } from '@/types/main';
 import { useAPIInfinite } from '@/utils/fetcher';
@@ -59,24 +58,19 @@ function ArchivedPage() {
   const { t } = useTranslation('translation', { keyPrefix: 'pages.archived' });
 
   return (
-    <>
-      {/* 功能型页面：使用站点默认信息 */}
-      <PageMeta />
-
-      <ContainerWithSideBar
-        sidebar={<SideBar />}
-        sidebarHeader={
-          <div className="flex items-center gap-2 p-4 text-lg font-semibold">
-            <div className="flex items-center gap-2">
-              <ChartAreaIcon className="size-5" />
-              {t('statistics')}
-            </div>
+    <ContainerWithSideBar
+      sidebar={<SideBar />}
+      sidebarHeader={
+        <div className="flex items-center gap-2 p-4 text-lg font-semibold">
+          <div className="flex items-center gap-2">
+            <ChartAreaIcon className="size-5" />
+            {t('statistics')}
           </div>
-        }
-      >
-        <MainPage />
-      </ContainerWithSideBar>
-    </>
+        </div>
+      }
+    >
+      <MainPage />
+    </ContainerWithSideBar>
   );
 }
